added hideable component and am now using that in donation form. run time issue of too many rerenders
