+++ components/DonationForm/DonationForm.tsx
@@ -1,43 +1,41 @@
 import React, { useState, Dispatch, SetStateAction } from 'react';
 import { Button, Platform } from 'react-native';
-
-import DateTimePicker from '@react-native-community/datetimepicker';
 import { Input } from 'react-native-elements';
-
+import HidableDatePicker from './HideableDatePicker';
 import { Text, View } from '../Themed';
 
-function HidableDatePicker(props: {
-  datetime: Date,
-  setDatetime: Dispatch<SetStateAction<Date>>
-}) {
-  const [showPicker, setShowPicker] = useState(false);
+function DonationForm() {
   const [description, setDescription] = useState('');
   const [pickupInstructions, setPickupInstructions] = useState('');
   const [weight, setWeight] = useState(0);
   const [error, setError] = useState('');
+  const [startDatetime, setStartDatetime] = useState(new Date());
 
-  const onDatetimeChange = (event: any, selectedDatetime?: Date) => {
-    const currentDatetime = selectedDatetime || props.datetime;
+  // // Initially, the start datetime will be now, and the end will be a day from now
+  const [endDatetime, setEndDatetime] = useState(new Date(Date.now() + 60 * 60 * 24 * 1000));
+  const currentDatetime = startDatetime;
 
-    // Workaround for Android issue: https://github.com/react-native-datetimepicker/datetimepicker/issues/54
-    setShowPicker(Platform.OS === 'ios');
+  // const [showStartDate, setShowStartDate] = useState(false);
+  // const [showEndDate, setShowEndDate] = useState(false);
 
-    props.setDatetime(selectedDatetime || props.datetime);
-  };
+  // const [description, setDescription] = useState('');
+  // const [pickupInstructions, setPickupInstructions] = useState('');
+  // const [weight, setWeight] = useState(0);
 
-  const toggleShowPicker = (event: any) => {
-    setShowPicker((shown) => !shown);
-  };
+  // const onStartDatetimeChange = (event: any, selectedDatetime?: Date) => {
+  //   const currentDatetime = selectedDatetime || startDatetime;
 
   // Date.now() and currentDatetime.getTime() return milliseconds
   // We check if the date is within a valid range - in this case,
   // at least three hours in the future
   const cutoff = Date.now() + 60 * 60 * 3 * 1000;
-  if (cutoff > currentDatetime.getTime() || cutoff > endDatetime.getTime()) {
+  if (cutoff > currentDatetime.getTime() || cutoff > startDatetime.getTime()) {
     setError('Date must be at least 3 hours from now!');
-  } else if (currentDatetime > endDatetime) {
+  } else if (startDatetime > currentDatetime) {
     setError('Start availability must be before end availability!');
   } else setError('');
+
+  setEndDatetime(currentDatetime);
 
   const handleSubmit = () => {
     fetch('http://localhost:3000/api/donations', {
@@ -62,51 +60,9 @@ function HidableDatePicker(props: {
     <View style={{ width: '100%' }}>
       <View style={{ width: '100%' }}>
         <Text>Availability Start</Text>
-        <Button
-          onPress={toggleShowPicker}
-          title={props.datetime.toString()}
-        />
-        {showPicker && (
-          <View>
-            {/* The date has to be in the code first otherwise the button title wont update properly */}
-            <DateTimePicker
-              value={props.datetime}
-              mode="date"
-              style={{ width: '100%' }}
-              onChange={onDatetimeChange}
-            />
-            <DateTimePicker
-              value={props.datetime}
-              mode="time"
-              style={{ width: '100%' }}
-              minuteInterval={1}
-              onChange={onDatetimeChange}
-            />
-          </View>
-        )}
+        <HidableDatePicker datetime={startDatetime} setDatetime={setStartDatetime} />
         <Text>Availability End</Text>
-        <Button
-          onPress={toggleShowPicker}
-          title={props.datetime.toString()}
-        />
-        {showPicker && (
-          <View>
-            {/* The date has to be in the code first otherwise the button title wont update properly */}
-            <DateTimePicker
-              value={props.datetime}
-              mode="date"
-              style={{ width: '100%' }}
-              onChange={onDatetimeChange}
-            />
-            <DateTimePicker
-              value={props.datetime}
-              mode="time"
-              style={{ width: '100%' }}
-              minuteInterval={1}
-              onChange={onDatetimeChange}
-            />
-          </View>
-        )}
+        <HidableDatePicker datetime={endDatetime} setDatetime={setEndDatetime} />
       </View>
       {/*
             Change to TextField at some point, or some other form of longer text input
@@ -140,4 +96,4 @@ function HidableDatePicker(props: {
   );
 }
 
-export default HidableDatePicker;
+export default DonationForm;
